Add access to database, fix dockerfiles

+++ BuildYourHead.Persistence/Entities/Product.cs
@@ -6,9 +6,9 @@
         public string Name { get; set; }
         public string Description { get; set; }
         public byte[] Image { get; set; }
-        public int Proteins { get; set; }
-        public int Carbohydrates { get; set; }
-        public int Fats { get; set; }
-        public int Nutrition { get; set; }
+        public double Proteins { get; set; }
+        public double Carbohydrates { get; set; }
+        public double Fats { get; set; }
+        public double Nutrition { get; set; }
     }
 }
